# server/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Starliner API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/deployments/databases": {
            "post": {
                "tags": [
                    "deployment"
                ],
                "summary": "Deploy database",
                "operationId": "deployDatabase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deploy Database",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeployDatabase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/git": {
            "post": {
                "tags": [
                    "deployment"
                ],
                "summary": "Deploy from Git Repository",
                "operationId": "deployFromGitRepository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deploy from Git",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeployFromGit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/git/{deploymentId}": {
            "put": {
                "tags": [
                    "deployment"
                ],
                "summary": "Update Deploy from Git",
                "operationId": "updateDeployFromGitRepository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Deploy from Git",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDeployFromGit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/images": {
            "post": {
                "tags": [
                    "deployment"
                ],
                "summary": "Deploy image",
                "operationId": "deployImage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deploy Image",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeployImage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/images/{deploymentId}": {
            "put": {
                "tags": [
                    "deployment"
                ],
                "summary": "Update image deployment",
                "operationId": "updateImageDeployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Image",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateImage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/ingresses": {
            "post": {
                "tags": [
                    "deployment"
                ],
                "summary": "Deploy ingress",
                "operationId": "deployIngress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deploy Ingress",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeployIngress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/ingresses/{deploymentId}": {
            "put": {
                "tags": [
                    "deployment"
                ],
                "summary": "Update ingress deployment",
                "operationId": "updateIngressDeployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Ingress",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateIngress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/{id}": {
            "delete": {
                "tags": [
                    "deployment"
                ],
                "summary": "Delete deployment",
                "operationId": "deleteDeployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/deployments/{id}/logs": {
            "get": {
                "tags": [
                    "deployment"
                ],
                "summary": "Stream deployment logs",
                "operationId": "streamDeploymentLogs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deployment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Cache-Control": {
                                "type": "string",
                                "description": "no-cache"
                            },
                            "Connection": {
                                "type": "string",
                                "description": "keep-alive"
                            },
                            "Content-Type": {
                                "type": "string",
                                "description": "text/event-stream"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Arg": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "request.DeployDatabase": {
            "type": "object",
            "required": [
                "environmentId",
                "serviceName"
            ],
            "properties": {
                "environmentId": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
        "request.DeployFromGit": {
            "type": "object",
            "required": [
                "dockerfilePath",
                "environmentId",
                "envs",
                "gitUrl",
                "port",
                "projectRepositoryPath",
                "serviceName"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.Arg"
                    }
                },
                "dockerfilePath": {
                    "type": "string"
                },
                "environmentId": {
                    "type": "integer"
                },
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EnvVar"
                    }
                },
                "gitUrl": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "projectRepositoryPath": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
        "request.DeployImage": {
            "type": "object",
            "required": [
                "environmentId",
                "envs",
                "imageName",
                "port",
                "serviceName",
                "tag"
            ],
            "properties": {
                "environmentId": {
                    "type": "integer"
                },
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EnvVar"
                    }
                },
                "imageName": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "volumeMountPath": {
                    "type": "string"
                },
                "volumeSizeMiB": {
                    "type": "integer"
                }
            }
        },
        "request.DeployIngress": {
            "type": "object",
            "required": [
                "environmentId",
                "ingressHosts"
            ],
            "properties": {
                "environmentId": {
                    "type": "integer"
                },
                "ingressHosts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.IngressHost"
                    }
                }
            }
        },
        "request.EnvVar": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "request.IngressHost": {
            "type": "object",
            "required": [
                "host",
                "paths"
            ],
            "properties": {
                "host": {
                    "type": "string"
                },
                "paths": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.IngressPath"
                    }
                }
            }
        },
        "request.IngressPath": {
            "type": "object",
            "required": [
                "path",
                "pathType",
                "serviceName"
            ],
            "properties": {
                "path": {
                    "type": "string"
                },
                "pathType": {
                    "type": "string",
                    "enum": [
                        "Prefix",
                        "Exact"
                    ]
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateDeployFromGit": {
            "type": "object",
            "required": [
                "dockerfilePath",
                "environmentId",
                "envs",
                "port",
                "projectRepositoryPath"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.Arg"
                    }
                },
                "dockerfilePath": {
                    "type": "string"
                },
                "environmentId": {
                    "type": "integer"
                },
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EnvVar"
                    }
                },
                "port": {
                    "type": "integer"
                },
                "projectRepositoryPath": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateImage": {
            "type": "object",
            "required": [
                "environmentId",
                "envs",
                "imageName",
                "port",
                "tag"
            ],
            "properties": {
                "environmentId": {
                    "type": "integer"
                },
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EnvVar"
                    }
                },
                "imageName": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "request.UpdateIngress": {
            "type": "object",
            "required": [
                "environmentId",
                "ingressHosts"
            ],
            "properties": {
                "environmentId": {
                    "type": "integer"
                },
                "ingressHosts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.IngressHost"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}